+++ lib/src/models/flutter_version_model.dart
@@ -116,11 +116,11 @@ class FlutterVersion {
   String get printFriendlyName {
     // Uppercase
 
-    if (isChannel) return 'Channel: ${uppercase(name)}';
+    if (isChannel) return 'Channel : ${name.capitalize}';
 
-    if (isCommit) return 'Commit: $name';
+    if (isCommit) return 'Commit : $name';
 
-    return 'SDK Version: $name';
+    return 'SDK Version : $name';
   }
 
   /// Compares CacheVersion with [other]