+++ lib/src/utils/io_utils.dart
@@ -25,5 +25,4 @@ void createLink(Link source, Directory target) {
     target.path,
     recursive: true,
   );
-  print('done');
 }